+++ src/danmaku_sender/service/__init__.py
@@ -1,0 +1,10 @@
+"""
+Service 层 — 业务逻辑
+
+包含无状态工具（parser、validator、exporter）和有状态引擎（scheduler、executor、monitor、session）。
+"""
+
+from .sender import DanmakuScheduler, DanmakuExecutor, SendingContext, SendJob
+
+
+__all__ = ["DanmakuScheduler", "DanmakuExecutor", "SendingContext", "SendJob"]
