+++ src/danmaku_sender/ui/controllers/sender_controller.py
@@ -7,15 +7,15 @@ from PySide6.QtCore import QObject, Signal, Slot
 from danmaku_sender.ui.framework.concurrency import WorkerThread, PoolTask
 
 from danmaku_sender.repo.history_manager import HistoryManager
-from danmaku_sender.core.engines.sender import DanmakuScheduler, DanmakuExecutor, SendingContext, SendJob
-from danmaku_sender.core.engines.sender.delay_manager import DelayManager
+from danmaku_sender.service import DanmakuScheduler, DanmakuExecutor, SendingContext, SendJob
+from danmaku_sender.service.sender.delay_manager import DelayManager
 from danmaku_sender.types.models.danmaku import Danmaku
 from danmaku_sender.types.models.result import DanmakuSendResult
 from danmaku_sender.types.models.common import VideoTarget, UnsentDanmakusRecord
 from danmaku_sender.config import ApiAuthConfig, SenderConfig
 from danmaku_sender.runtime.app_state import AppState
-from danmaku_sender.core.services.danmaku_parser import DanmakuParser
-from danmaku_sender.core.services.danmaku_exporter import create_xml_from_danmakus
+from danmaku_sender.service.danmaku_parser import DanmakuParser
+from danmaku_sender.service.danmaku_exporter import create_xml_from_danmakus
 from danmaku_sender.repo.bili_api_client import BiliApiClient
 from danmaku_sender.utils.system_utils import KeepSystemAwake
