+++ src/danmaku_sender/ui/controllers/video_controller.py
@@ -5,7 +5,7 @@ from PySide6.QtCore import QObject, Signal, QThreadPool, Slot
 from danmaku_sender.repo.bili_api_client import BiliApiClient
 from danmaku_sender.types.models.video import VideoInfo
 from danmaku_sender.config import ApiAuthConfig
-from danmaku_sender.core.services.video_fetcher import VideoFetcher
+from danmaku_sender.service.video_fetcher import VideoFetcher
 from danmaku_sender.ui.framework.concurrency import PoolTask
